+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ## Version 1.27
 
+- Add option to configure default popup tab [feature 877](https://github.com/tprouvot/Salesforce-Inspector-reloaded/issues/877) (requested by William Salomon)
 - Support Manual channel configuration and Change Data Capture Events on `Event Monitor` [feature 859](https://github.com/tprouvot/Salesforce-Inspector-reloaded/issues/859) (idea by [David Moruzzi](https://github.com/dmoruzzi)) and [feature 867](https://github.com/tprouvot/Salesforce-Inspector-reloaded/issues/867) (idea by Ammon Ho)
 - `Metadata` Handle SpecifiedTest + Start zip download when ready [feature 319](https://github.com/tprouvot/Salesforce-Inspector-reloaded/issues/319) request by [Alfredo Chissotti](https://github.com/Astisme)
 - Add Metadata Deploy feature